Remove prioritization_fee.rs
