+++ compute-budget/src/lib.rs
@@ -3,4 +3,3 @@
 
 pub mod compute_budget;
 pub mod compute_budget_limits;
-pub mod prioritization_fee;
